requirements.txt
flask==3.1.1
nltk==3.9.2
opensearch-py==3.0.0
psycopg2-binary
pypdf==6.6.2
python-dotenv==1.2.1
sentence-transformers==5.1.2
tiktoken==0.12.0
